Exception Handling refactor
